Rename global to typings
